Bug fix: bad msg.sender usage
`msg.sender` would get applied to the original minter and token owner. Since the ERC20 contract is the one making the function call, the certificate would be owned by the contract. This commit ensures the right owner address is used in the process. The flaw was introduced when migrating minting controls from the ERC721 contract to the ERC20.